mysite/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
Markdown==3.7
mysqlclient==2.2.7
packaging==24.2
python-decouple==3.8
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
